Caching thumbnail images now

+++ Utils/TempFolder.cs
@@ -1,13 +1,12 @@
 using System;
 using System.IO;
+using System.Text;
 
 namespace VideoCompressorGUI.Utils;
 
 public static class TempFolder
 {
     private static string ROOT = "";
-    private static uint counter = 0;
-    private static object mutex = new object();
     
     static TempFolder()
     {
@@ -15,29 +14,44 @@ public static class TempFolder
         Directory.CreateDirectory(ROOT);
     }
     
-    public static void Clear()
+    public static void ClearOnTimeExpired()
     {
         // Clear folder
         DirectoryInfo di = new DirectoryInfo(ROOT);
-        
+
+        bool cleanup = false;
         foreach (FileInfo file in di.GetFiles())
-            file.Delete();
+        {
+            if (DateTime.Now - file.CreationTime > TimeSpan.FromSeconds(60)/*TimeSpan.FromDays(30)*/)
+            {
+                cleanup = true;
+                Console.WriteLine("Clean up temp folder");
+                break;
+            }
+        }
+
+        if (cleanup)
+        {
+            // Clear folder
+            foreach (FileInfo file in di.GetFiles())
+                file.Delete();
         
-        foreach (DirectoryInfo directoryInfo in di.GetDirectories())
-            directoryInfo.Delete(true);
+            foreach (DirectoryInfo directoryInfo in di.GetDirectories())
+                directoryInfo.Delete(true);
+        }
     }
-    
-    public static string GenerateNewName(string extension)
+
+    public static string GenerateThumbnailName(in DateTime creationTime, string filePath, string extension)
     {
+        var creationString = creationTime.ToShortDateString();
+        var fileName = Path.GetFileNameWithoutExtension(filePath);
+        
         // Creates if needed
         Directory.CreateDirectory(ROOT);
 
-        lock (mutex)
-        {
-            counter++;
-        }
+        StringBuilder builder = new StringBuilder(ROOT);
+        builder.Append("/").Append(fileName).Append('-').Append(creationString).Append(extension);
         
-        string base64Guid = "_" + counter; //Convert.ToBase64String(Guid.NewGuid().ToByteArray());
-        return ROOT + "/" + base64Guid + extension;
+        return builder.ToString();
     }
 }
